refactor: massive refactor of project
Moved everything into a Class for better management of variables and methods. Works with threading, though the cleanup needs to be improved. Maybe introduce the possibility of sending a notification to the user when the AutoClicker is started and stopped. The program can only be killed from the terminal currently, which is not ideal, so that needs to be attended to in the future, will create an issue for this soon. Next commit will be documentation.